Move Temporalio classes to Temporalio.Common
