+++ src/Temporalio/Client/Schedules/ScheduleOptions.cs
@@ -1,5 +1,6 @@
 using System;
 using System.Collections.Generic;
+using Temporalio.Common;
 
 namespace Temporalio.Client.Schedules
 {
