+++ src/Temporalio/Client/WorkflowHandle.cs
@@ -10,6 +10,7 @@ using Temporalio.Exceptions;
 #if NETCOREAPP3_0_OR_GREATER
 using System.Runtime.CompilerServices;
 using System.Threading;
+using Temporalio.Common;
 #endif
 
 namespace Temporalio.Client
