+++ src/Temporalio/Workflows/IWorkflowContext.cs
@@ -3,6 +3,7 @@ using System.Collections.Generic;
 using System.Threading;
 using System.Threading.Tasks;
 using Microsoft.Extensions.Logging;
+using Temporalio.Common;
 using Temporalio.Converters;
 
 namespace Temporalio.Workflows
