+++ src/Temporalio/Workflows/LocalActivityOptions.cs
@@ -1,5 +1,6 @@
 using System;
 using System.Threading;
+using Temporalio.Common;
 
 namespace Temporalio.Workflows
 {
