+++ src/Temporalio/Workflows/Workflow.cs
@@ -6,6 +6,7 @@ using System.Linq.Expressions;
 using System.Threading;
 using System.Threading.Tasks;
 using Microsoft.Extensions.Logging;
+using Temporalio.Common;
 using Temporalio.Converters;
 
 namespace Temporalio.Workflows
