+++ tests/Temporalio.Tests/Worker/WorkflowWorkerTests.cs
@@ -9,6 +9,7 @@ using Temporalio.Activities;
 using Temporalio.Api.Enums.V1;
 using Temporalio.Api.History.V1;
 using Temporalio.Client;
+using Temporalio.Common;
 using Temporalio.Converters;
 using Temporalio.Exceptions;
 using Temporalio.Worker;
